fix: type Todo
